==> part-3-descriptor-calculation.ipynb <==
We will be building a machine learning model using the ChEMBL bioactivity data. We will be calculating molecular descriptors that are essentailly quantitative description of the compunds in the dataset. Finally, we will be preparing this into a dataset for subsequent model building.

## Download PaDEL-Descriptor

! wget https://github.com/dataprofessor/bioinformatics/raw/master/padel.zip
! wget https://github.com/dataprofessor/bioinformatics/raw/master/padel.sh

! unzip padel.zip

## Load bioactivity data

Download the curated ChEMBL bioactivity data that has been pre-processed from earlier parts. Here we will be using the bioactivity_data_3class_pIC50.csv file that essentially contain the pIC50 values that we will be using for building a regression model.

! wget https://raw.githubusercontent.com/dataprofessor/data/master/acetylcholinesterase_04_bioactivity_data_3class_pIC50.csv

import pandas as pd

df3 = pd.read_csv("acetylcholinesterase_04_bioactivity_data_3class_pIC50.csv")

df3

selection = ["canonical_smiles", "molecule_chembl_id"]
df3_selection = df3[selection]
df3_selection.to_csv("molecule.smi", sep="\t", index=False, header=False)

! cat molecule.smi | head -5

# How many lines of molecules do we have
! cat molecule.smi | wc -l

## Calculate fingerprint descriptors

### Calculate PaDEL descriptors

! bash padel.sh

## Preparing the X and Y data matrices

### X data matrix

df3_x = pd.read_csv("descriptors_output.csv")

df3_x

df3_x = df3_x.drop(columns=["Name"])
df3_x

### Y data matrix

### Convert IC50 to pIC50

df3_y = df3["pIC50"]
df3_y

## Combining X and Y variable

dataset3 = pd.concat([df3_x, df3_y], axis=1)
dataset3

dataset3.to_csv("acetylcholinesterase_06_bioactivity_data_3class_pIC50_pubchem_fp.csv", index=False)
