==> part-4-model-building.ipynb <==
We will be building a regression model of acetylocholinesterase inhibtors using the random forest algorithm.

### Import libraries

import pandas as pd
import seaborn as sns
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor

### Load the dataset

df = pd.read_csv("acetylcholinesterase_06_bioactivity_data_3class_pIC50_pubchem_fp.csv")

df

### Input features

The Acetyloholinesterase dataset contains 881 input features and 1 output variable (pIC50 values).

x = df.drop("pIC50", axis=1)

x

### Output features

y = df.pIC50

y

### Examine the data dimension

x.shape

y.shape

### Remove low variance features

from sklearn.feature_selection import VarianceThreshold

selection = VarianceThreshold(threshold=(.8 * (1 - .8)))
x = selection.fit_transform(x)

x.shape

### Data split (80/20 ratio)

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=0.2)

x_train.shape, y_train.shape

x_test.shape, y_test.shape

### Building a regression model using random forest

import numpy as np
np.random.seed(100)

model = RandomForestRegressor(n_estimators=100)
model.fit(x_train, y_train)
r2 = model.score(x_test, y_test)

r2

y_pred = model.predict(x_test)

### Scatter plot of experimental vs predicted pIC50 values

import seaborn as sns
import matplotlib.pyplot as plt

sns.set(color_codes=True)
sns.set_style("white")

ax = sns.regplot(y_test, y_pred, scatter_kws={"alpha": 0.4})
ax.set_xlabel("Experimental pIC50", fontsize="large", fontweight="bold")
ax.set_ylabel("Predicted pIC50", fontsize="large", fontweight="bold")
ax.set_xlim(0, 12)
ax.set_ylim(0, 12)
ax.figure.set_size_inches(5, 5)

plt.show
